Deconstruct Cypher show queries
