+++ src/neo4j/cypher-queries/character/show/index.js
@@ -1,0 +1,7 @@
+import getShowMaterialsQuery from './show-materials';
+import getShowProductionsQuery from './show-productions';
+
+export default () => [
+	getShowMaterialsQuery(),
+	getShowProductionsQuery()
+];
