+++ src/neo4j/cypher-queries/material/show/show-awards.js
@@ -1,0 +1,267 @@
+export default () => `
+	MATCH (material:Material { uuid: $uuid })
+
+	OPTIONAL MATCH (material)-[:HAS_SUB_MATERIAL*0..1]-(materialLinkedToCategory:Material)
+		<-[nomineeRel:HAS_NOMINEE]-(category:AwardCeremonyCategory)
+		<-[categoryRel:PRESENTS_CATEGORY]-(ceremony:AwardCeremony)
+
+	OPTIONAL MATCH (ceremony)<-[:PRESENTED_AT]-(award:Award)
+
+	OPTIONAL MATCH (category)-[nominatedEntityRel:HAS_NOMINEE]->(nominatedEntity)
+		WHERE
+			(
+				(nominatedEntity:Person AND nominatedEntityRel.nominatedCompanyUuid IS NULL) OR
+				nominatedEntity:Company
+			) AND
+			(
+				nomineeRel.nominationPosition IS NULL OR
+				nomineeRel.nominationPosition = nominatedEntityRel.nominationPosition
+			)
+
+	WITH
+		material,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		nominatedEntityRel,
+		CASE WHEN material <> materialLinkedToCategory
+			THEN materialLinkedToCategory
+			ELSE null
+		END AS recipientMaterial,
+		COLLECT(nominatedEntity {
+			model: TOUPPER(HEAD(LABELS(nominatedEntity))),
+			.uuid,
+			.name,
+			nominatedMemberUuids: nominatedEntityRel.nominatedMemberUuids
+		}) AS nominatedEntities
+
+	UNWIND (CASE nominatedEntities WHEN [] THEN [null] ELSE nominatedEntities END) AS nominatedEntity
+
+		UNWIND (COALESCE(nominatedEntity.nominatedMemberUuids, [null])) AS nominatedMemberUuid
+
+			OPTIONAL MATCH (category)-[nominatedMemberRel:HAS_NOMINEE]->
+				(nominatedMember:Person { uuid: nominatedMemberUuid })
+				WHERE
+					nominatedEntityRel.nominationPosition IS NULL OR
+					nominatedEntityRel.nominationPosition = nominatedMemberRel.nominationPosition
+
+			WITH
+				material,
+				recipientMaterial,
+				nomineeRel,
+				category,
+				categoryRel,
+				ceremony,
+				award,
+				nominatedEntityRel,
+				nominatedEntity,
+				nominatedMember
+				ORDER BY nominatedMemberRel.memberPosition
+
+			WITH
+				material,
+				recipientMaterial,
+				nomineeRel,
+				category,
+				categoryRel,
+				ceremony,
+				award,
+				nominatedEntityRel,
+				nominatedEntity,
+				COLLECT(nominatedMember { model: 'PERSON', .uuid, .name }) AS nominatedMembers
+
+	WITH
+		material,
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		nominatedEntityRel,
+		nominatedEntity,
+		nominatedMembers
+		ORDER BY nominatedEntityRel.nominationPosition, nominatedEntityRel.entityPosition
+
+	WITH
+		material,
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		COLLECT(
+			CASE nominatedEntity WHEN NULL
+				THEN null
+				ELSE nominatedEntity { .model, .uuid, .name, members: nominatedMembers }
+			END
+		) AS nominatedEntities
+
+	WITH
+		material,
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		[nominatedEntity IN nominatedEntities | CASE nominatedEntity.model WHEN 'COMPANY'
+			THEN nominatedEntity
+			ELSE nominatedEntity { .model, .uuid, .name }
+		END] AS nominatedEntities
+
+	OPTIONAL MATCH (category)-[nominatedProductionRel:HAS_NOMINEE]->(nominatedProduction:Production)
+		WHERE
+			(
+				nomineeRel.nominationPosition IS NULL OR
+				nomineeRel.nominationPosition = nominatedProductionRel.nominationPosition
+			)
+
+	OPTIONAL MATCH (nominatedProduction)-[:PLAYS_AT]->(venue:Venue)
+
+	OPTIONAL MATCH (venue)<-[:HAS_SUB_VENUE]-(surVenue:Venue)
+
+	OPTIONAL MATCH (nominatedProduction)<-[:HAS_SUB_PRODUCTION]-(surProduction:Production)
+
+	WITH
+		material,
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		nominatedEntities,
+		nominatedProductionRel,
+		nominatedProduction,
+		venue,
+		surVenue,
+		surProduction
+		ORDER BY nominatedProductionRel.productionPosition
+
+	WITH
+		material,
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		nominatedEntities,
+		COLLECT(
+			CASE nominatedProduction WHEN NULL
+				THEN null
+				ELSE nominatedProduction {
+					model: 'PRODUCTION',
+					.uuid,
+					.name,
+					.startDate,
+					.endDate,
+					venue: CASE venue WHEN NULL
+						THEN null
+						ELSE venue {
+							model: 'VENUE',
+							.uuid,
+							.name,
+							surVenue: CASE surVenue WHEN NULL
+								THEN null
+								ELSE surVenue { model: 'VENUE', .uuid, .name }
+							END
+						}
+					END,
+					surProduction: CASE surProduction WHEN NULL
+						THEN null
+						ELSE surProduction { model: 'PRODUCTION', .uuid, .name }
+					END
+				}
+			END
+		) AS nominatedProductions
+
+	OPTIONAL MATCH (category)-[coNominatedMaterialRel:HAS_NOMINEE]->(coNominatedMaterial:Material)
+		WHERE
+			(
+				nomineeRel.nominationPosition IS NULL OR
+				nomineeRel.nominationPosition = coNominatedMaterialRel.nominationPosition
+			) AND
+			coNominatedMaterial.uuid <> material.uuid AND
+			NOT EXISTS((material)-[:HAS_SUB_MATERIAL]-(coNominatedMaterial))
+
+	OPTIONAL MATCH (coNominatedMaterial)<-[:HAS_SUB_MATERIAL]-(coNominatedMaterialSurMaterial:Material)
+
+	WITH
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		nominatedEntities,
+		nominatedProductions,
+		coNominatedMaterialRel,
+		coNominatedMaterial,
+		coNominatedMaterialSurMaterial
+		ORDER BY coNominatedMaterialRel.materialPosition
+
+	WITH
+		recipientMaterial,
+		nomineeRel,
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		nominatedEntities,
+		nominatedProductions,
+		COLLECT(
+			CASE coNominatedMaterial WHEN NULL
+				THEN null
+				ELSE coNominatedMaterial {
+					model: 'MATERIAL',
+					.uuid,
+					.name,
+					.format,
+					.year,
+					surMaterial: CASE coNominatedMaterialSurMaterial WHEN NULL
+						THEN null
+						ELSE coNominatedMaterialSurMaterial { model: 'MATERIAL', .uuid, .name }
+					END
+				}
+			END
+		) AS coNominatedMaterials
+		ORDER BY nomineeRel.nominationPosition
+
+	WITH
+		category,
+		categoryRel,
+		ceremony,
+		award,
+		COLLECT({
+			model: 'NOMINATION',
+			isWinner: COALESCE(nomineeRel.isWinner, false),
+			type: COALESCE(nomineeRel.customType, CASE WHEN nomineeRel.isWinner THEN 'Winner' ELSE 'Nomination' END),
+			recipientMaterial: CASE recipientMaterial WHEN NULL
+				THEN null
+				ELSE recipientMaterial { model: 'MATERIAL', .uuid, .name, .format, .year }
+			END,
+			entities: nominatedEntities,
+			productions: nominatedProductions,
+			coMaterials: coNominatedMaterials
+		}) AS nominations
+		ORDER BY categoryRel.position
+
+	WITH
+		ceremony,
+		award,
+		COLLECT(category { model: 'AWARD_CEREMONY_CATEGORY', .name, nominations }) AS categories
+		ORDER BY ceremony.name DESC
+
+	WITH
+		award,
+		COLLECT(ceremony { model: 'AWARD_CEREMONY', .uuid, .name, categories }) AS ceremonies
+		ORDER BY award.name
+
+	RETURN
+		COLLECT(award { model: 'AWARD', .uuid, .name, ceremonies }) AS awards
+`;
