+++ src/neo4j/cypher-queries/production/show/index.js
@@ -1,0 +1,7 @@
+import getShowQuery from './show';
+import getShowAwardsQuery from './show-awards';
+
+export default () => [
+	getShowQuery(),
+	getShowAwardsQuery()
+];
